% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchFlanks.R
\name{FetchFlanks}
\alias{FetchFlanks}
\title{FetchFlanks()}
\usage{
FetchFlanks(
  vcf_list = "No Default",
  Reference_Genome = "No Default",
  validation = TRUE,
  Flank_Length = 75,
  Output_Base_Name = "No Default"
)
}
\arguments{
\item{vcf_list}{R object into which vcf data was stored with ReadVcf()}

\item{Reference_Genome}{The reference genome unto which the sequence data has been mapped (including .fasta or .fna file extension).}

\item{validation}{Whether or not the submitted reference genome is compared to data from the vcf_list to verify that the correct reference has been submitted. This prevents the function from potentially generating flanking regions for non-existing alleles.}

\item{Flank_Length}{How many base pairs left and right of the variant are to be extracted. The default is 75 bp which produces a sequence of 150 bp (75 left of the variant and 75 right of the variant).}

\item{Output_Base_Name}{To save the flanking regions directly to an excel file, insert the desired base file name here. The file extension is automatically added to this name.}
}
\description{
This function is used to retrieve the sequences surrounding the variants for primer construction.
}
\examples{
# Retrieve a short region of 10 bp and save it to an excel file:
FetchFlanks(vcf_list=vcf_object,Reference_Genome="MyRef.fasta",Flank_Length=10,Output_Base_Name="Regions"))
}
\keyword{construction}
\keyword{flanking}
\keyword{flanks}
\keyword{primer}
\keyword{region}
\keyword{vcf}
\keyword{vcfR}
