% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scan_Vcf.R
\name{Scan_Vcf}
\alias{Scan_Vcf}
\title{Scan_Vcf()}
\usage{
Scan_Vcf(vcf_file = "No Default")
}
\arguments{
\item{vcf_file}{The Variant Call Format (VCF) file name to scan (including .vcf extension). This function is incompatible with .gz compressed file.}
}
\description{
Based on VariantAnnotation, this function allows the user to find the starting row and number of variants of all chromosomes in a vcf file. This function is meant to provide an overview of the number of variants called, as well as the nrows and skip arguments required when reading in a vcf per chromosome.
}
\examples{
# Scan a vcf file:
Scan_Vcf("file.vcf")
}
\keyword{exploration}
\keyword{scan}
\keyword{vcf}
